requirements.txt
aiofiles==24.1.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
async-timeout==4.0.2
attrs==24.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.5.0
cdp-socket==1.2.8
certifi==2023.7.22
cffi==1.17.0
charset-normalizer==3.2.0
colorama==0.4.6
dateparser==1.2.0
distro==1.8.0
exceptiongroup==1.2.0
frozenlist==1.4.0
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-search-results==2.4.2
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.4
jiter==0.8.0
multidict==6.0.4
numpy==2.1.0
oauthlib==3.2.2
openai==1.55.2
orjson==3.10.7
outcome==1.3.0.post0
platformdirs==4.2.2
praw==7.7.1
prawcore==2.3.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
regex==2024.7.24
requests==2.28.2
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.14.1
selenium==4.24.0
selenium_driverless==1.9.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.66.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.16
websocket-client==1.8.0
websockets==13.0.1
wsproto==1.2.0
yarl==1.9.2
